+++ src/Multiphase/fluid_particle_interaction/Intersections_Face_Facettes_Data.h
@@ -36,10 +36,10 @@ public: // On copie ce qui est fait dans Intersections_Elem_Facettes_Data (parce
   // En 2D, on a barycentre_[2] = 0.;
   double barycentre_[3];
 
-  long index_face_suivante_;  // -1 si derniere face de la liste
-  long index_facette_suivante_; // idem.
-  long numero_facette_;
-  long numero_face_;
+  int index_face_suivante_;  // -1 si derniere face de la liste
+  int index_facette_suivante_; // idem.
+  int numero_facette_;
+  int numero_face_;
 };
 
 // ====================================================================
@@ -50,7 +50,7 @@ public: // On copie ce qui est fait dans Intersections_Elem_Facettes_Data (parce
 // doublement chainee.
 // Pour parcourir les facettes qui coupent une face "face", on fait:
 //
-//   long index=index_face()[face];
+//   int index=index_face()[face];
 //   while (index >= 0) {
 //     const Intersections_Face_Facettes_Data & data = data_intersection(index);
 //     ... // faire quelque chose avec data
@@ -59,7 +59,7 @@ public: // On copie ce qui est fait dans Intersections_Elem_Facettes_Data (parce
 //
 // Pour parcourir les faces qui sont coupes par une facette "facette":
 //
-//   long index=index_facette()[facette];
+//   int index=index_facette()[facette];
 //   while (index >= 0) {
 //     const Intersections_Face_Facettes_Data & data = data_intersection(index);
 //     ... // faire quelque chose avec data
@@ -71,12 +71,12 @@ class Intersections_Face_Facettes
 public:
   Intersections_Face_Facettes();
   ~Intersections_Face_Facettes();
-  void get_liste_faces_traversees(long num_facette,
+  void get_liste_faces_traversees(int num_facette,
                                   ArrOfInt& liste_faces) const;
-  void get_liste_facettes_traversantes(long num_face,
+  void get_liste_facettes_traversantes(int num_face,
                                        ArrOfInt& liste_facettes) const;
-  void ajoute_intersection(long num_facette,
-                           long num_face,
+  void ajoute_intersection(int num_facette,
+                           int num_face,
                            double surface_intersection,
                            double contrib_volume_phase1,
                            double barycentre_u,
@@ -87,12 +87,12 @@ public:
   // nb_faces_euleriennes doit etre correct.
   // nb_facettes peut etre une estimation seulement (en pratique lors du
   // parcours on ne peut pas prevoir le nombre final de facettes)
-  void reset(long nb_faces_euleriennes=0, long nb_facettes=0);
+  void reset(int nb_faces_euleriennes=0, int nb_facettes=0);
 
   const ArrOfInt& index_face() const;
   const ArrOfInt& index_facette() const;
-  inline const Intersections_Face_Facettes_Data& data_intersection(long index) const;
-  inline Intersections_Face_Facettes_Data& get_set_data_intersection(long index);
+  inline const Intersections_Face_Facettes_Data& data_intersection(int index) const;
+  inline Intersections_Face_Facettes_Data& get_set_data_intersection(int index);
 
   //operateur de copie
   const Intersections_Face_Facettes& operator=(const Intersections_Face_Facettes& iff);
@@ -110,15 +110,15 @@ private:
   //
   // Les donnees : une face par couple facette/face traverse.
   // (politique d'allocation : on agrandit d'un facteur 2, on ne diminue jamais).
-  long data_allocated_size; // Le nombre de faces allouees
-  long data_real_size;      // Le nombre de faces reellement utilisees
+  int data_allocated_size; // Le nombre de faces allouees
+  int data_real_size;      // Le nombre de faces reellement utilisees
   Intersections_Face_Facettes_Data * data;
 };
 
 // Description:
 //  Renvoie les donnees de l'intersection stockee a l'indice "index"
 //  dans le tableau "data" ( 0 <= index < data_real_size_ )
-inline const Intersections_Face_Facettes_Data& Intersections_Face_Facettes::data_intersection(long index) const
+inline const Intersections_Face_Facettes_Data& Intersections_Face_Facettes::data_intersection(int index) const
 {
   //Cerr << "data_real_size " << data_real_size << finl;
   assert(index >= 0 && index < data_real_size);
@@ -128,7 +128,7 @@ inline const Intersections_Face_Facettes_Data& Intersections_Face_Facettes::data
 //  Renvoie les donnees de l'intersection stockee a l'indice "index"
 //  dans le tableau "data" ( 0 <= index < data_real_size_ )
 // ATTENTION A SON UTILISATION !!!
-inline Intersections_Face_Facettes_Data& Intersections_Face_Facettes::get_set_data_intersection(long index)
+inline Intersections_Face_Facettes_Data& Intersections_Face_Facettes::get_set_data_intersection(int index)
 {
   assert(index >= 0 && index < data_real_size);
   return data[index];
